I have added +/- buttons in recipe cart

+++ app/controllers/concerns/current_cart.rb
@@ -10,8 +10,10 @@ module CurrentCart
 				session[:cart_id] = @cart.id
 
 				@old = Cart.where("user_id = ? and id != ?", current_user.id, @cart).first
-				Position.where("cart_id = ?", @old).each do |pos| # удаление всех связанных с предыдущей корзиной рецептов
-					pos.destroy
+				if @old
+					Position.where("cart_id = ?", @old).each do |pos| # удаление всех связанных с предыдущей корзиной рецептов
+						pos.destroy
+					end
 				end
 				@old.destroy # удаление предыдущей версии корзины
 			end
